services/garuda_api/pyproject.toml
[project]
name = "garuda-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-jose[cryptography]",
    "requests",
    "python-dotenv",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "pydantic-settings"
]
